requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
Django==5.2
django-tailwind==4.0.1
django-unfold==0.52.0
djangorestframework==3.16.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.7
numpy==2.2.4
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.2.1
Pygments==2.19.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
text-unidecode==1.3
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
